requirements.txt
transformers
torch
pillow
pathlib
